Added about activity

+++ app/src/main/java/com/vsdeni/ejru/helpers/Utils.java
@@ -115,11 +115,11 @@ public class Utils {
      * @param context Context to get resources and device specific display metrics
      * @return A float value to represent px equivalent to dp depending on device density
      */
-    public static float convertDpToPixel(float dp, Context context) {
+    public static int convertDpToPixel(float dp, Context context) {
         Resources resources = context.getResources();
         DisplayMetrics metrics = resources.getDisplayMetrics();
         float px = dp * (metrics.densityDpi / 160f);
-        return px;
+        return (int) px;
     }
 
     /**
